Rpi stuff/rpi_requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-dht==4.0.6
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.2
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
paho-mqtt==2.1.0
pigpio==1.78
pyftdi==0.56.0
pyserial==3.5
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
RPLCD==1.3.1
smbus2==0.5.0
sysv-ipc==1.1.0
typing_extensions==4.12.2
